update event handler and services
