+++ event-watchers/substrate/src/signature_bridge_watcher.rs
@@ -14,17 +14,18 @@
 
 use std::sync::Arc;
 use tokio::sync::Mutex;
+
+use webb::substrate::subxt::config::SubstrateConfig;
 use webb::substrate::subxt::ext::sp_core::hashing::keccak_256;
+
 use webb::substrate::subxt::{self, dynamic::Value, OnlineClient};
 
 use webb::substrate::protocol_substrate_runtime::api::signature_bridge::calls::{ExecuteProposal,SetMaintainer};
-use webb_event_watcher_traits::substrate::{BlockNumberOf, SubstrateBridgeWatcher};
+use webb_event_watcher_traits::substrate::{SubstrateBridgeWatcher, EventHandler};
 use webb_event_watcher_traits::SubstrateEventWatcher;
 use webb_relayer_store::sled::{SledQueueKey,SledStore};
 use webb_relayer_store::{BridgeCommand, QueueStore};
-use webb::substrate::{
-    protocol_substrate_runtime,
-};
+
 use webb::substrate::protocol_substrate_runtime::api as RuntimeApi;
 use webb::evm::ethers::utils;
 use webb::substrate::protocol_substrate_runtime::api::signature_bridge;
@@ -37,44 +38,53 @@ use webb::substrate::protocol_substrate_runtime::api::runtime_types::sp_core::bo
 
 /// A SignatureBridge contract events & commands watcher.
 #[derive(Copy, Clone, Debug, Default)]
-pub struct SubstrateBridgeEventWatcher;
+pub struct SubstrateBridgeEventHandler;
 
 #[async_trait::async_trait]
-impl SubstrateEventWatcher for SubstrateBridgeEventWatcher {
-    const TAG: &'static str = "Substrate bridge pallet Watcher";
-
-    const PALLET_NAME: &'static str = "SignatureBridge";
-
-    type RuntimeConfig = subxt::SubstrateConfig;
-
-    type Client = OnlineClient<Self::RuntimeConfig>;
-
-    type Event = protocol_substrate_runtime::api::Event;
-
-    type FilteredEvent = signature_bridge::events::MaintainerSet;
+impl EventHandler<SubstrateConfig> for SubstrateBridgeEventHandler {
+    type Client = OnlineClient<SubstrateConfig>;
 
     type Store = SledStore;
 
-    async fn handle_event(
+    async fn can_handle_event(
+        &self,
+        events: subxt::events::Events<SubstrateConfig>,
+    ) -> webb_relayer_utils::Result<bool> {
+        let has_event =
+            events.has::<signature_bridge::events::MaintainerSet>()?;
+        Ok(has_event)
+    }
+
+    async fn handle_events(
         &self,
         _store: Arc<Self::Store>,
-        _api: Arc<Self::Client>,
-        (event, _block_number): (Self::FilteredEvent, BlockNumberOf<Self>),
+        _client: Arc<Self::Client>,
+        (events, _block_number): (subxt::events::Events<SubstrateConfig>, u64),
         _metrics: Arc<Mutex<metric::Metrics>>,
     ) -> webb_relayer_utils::Result<()> {
         // todo
         // if the ownership is transferred to the new owner, we need to
         // to check our txqueue and remove any pending tx that was trying to
         // do this transfer.
-        tracing::event!(
-            target: webb_relayer_utils::probe::TARGET,
-            tracing::Level::DEBUG,
-            kind = %webb_relayer_utils::probe::Kind::SignatureBridge,
-            call = "pallet_signature_bridge:: set_maintainer",
-            msg = "Maintainer set",
-            new_maintainer = ?event.new_maintainer,
-            old_maintainer = ?event.old_maintainer,
-        );
+        if !self.can_handle_event(events.clone()).await? {
+            return Ok(());
+        }
+        let maintainer_set_events = events
+            .find::<signature_bridge::events::MaintainerSet>()
+            .flatten()
+            .collect::<Vec<_>>();
+
+        for event in maintainer_set_events {
+            tracing::event!(
+                target: webb_relayer_utils::probe::TARGET,
+                tracing::Level::DEBUG,
+                kind = %webb_relayer_utils::probe::Kind::SignatureBridge,
+                call = "pallet_signature_bridge:: set_maintainer",
+                msg = "Maintainer set",
+                new_maintainer = ?event.new_maintainer,
+                old_maintainer = ?event.old_maintainer,
+            );
+        }
 
         // mark this event as processed.
         // let events_bytes = &event.encode();
@@ -84,14 +94,25 @@ impl SubstrateEventWatcher for SubstrateBridgeEventWatcher {
     }
 }
 
+/// A SignatureBridge contract events & commands watcher.
+#[derive(Copy, Clone, Debug, Default)]
+pub struct SubstrateBridgeEventWatcher;
+
+impl SubstrateEventWatcher<SubstrateConfig> for SubstrateBridgeEventWatcher {
+    const TAG: &'static str = "Substrate bridge pallet Watcher";
+    const PALLET_NAME: &'static str = "SignatureBridge";
+    type Client = OnlineClient<SubstrateConfig>;
+    type Store = SledStore;
+}
+
 #[async_trait::async_trait]
-impl SubstrateBridgeWatcher for SubstrateBridgeEventWatcher {
+impl SubstrateBridgeWatcher<SubstrateConfig> for SubstrateBridgeEventWatcher {
     #[tracing::instrument(skip_all)]
     async fn handle_cmd(
         &self,
         chain_id: u32,
         store: Arc<Self::Store>,
-        api: Arc<Self::Client>,
+        client: Arc<Self::Client>,
         cmd: BridgeCommand,
     ) -> webb_relayer_utils::Result<()> {
         use BridgeCommand::*;
@@ -101,7 +122,7 @@ impl SubstrateBridgeWatcher for SubstrateBridgeEventWatcher {
                 self.execute_proposal_with_signature(
                     chain_id,
                     store,
-                    api.clone(),
+                    client.clone(),
                     (data, signature),
                 )
                 .await?
@@ -114,7 +135,7 @@ impl SubstrateBridgeWatcher for SubstrateBridgeEventWatcher {
                 self.transfer_ownership_with_signature(
                     chain_id,
                     store,
-                    api.clone(),
+                    client.clone(),
                     (public_key, nonce, signature),
                 )
                 .await?
@@ -126,14 +147,14 @@ impl SubstrateBridgeWatcher for SubstrateBridgeEventWatcher {
 
 impl SubstrateBridgeEventWatcher
 where
-    Self: SubstrateBridgeWatcher,
+    Self: SubstrateBridgeWatcher<SubstrateConfig>,
 {
     #[tracing::instrument(skip_all)]
     async fn execute_proposal_with_signature(
         &self,
         chain_id: u32,
-        store: Arc<<Self as SubstrateEventWatcher>::Store>,
-        api: Arc<<Self as SubstrateEventWatcher>::Client>,
+        store: Arc<<Self as SubstrateEventWatcher<SubstrateConfig>>::Store>,
+        api: Arc<<Self as SubstrateEventWatcher<SubstrateConfig>>::Client>,
         (proposal_data, signature): (Vec<u8>, Vec<u8>),
     ) -> webb_relayer_utils::Result<()> {
         let proposal_data_hex = hex::encode(&proposal_data);
@@ -229,8 +250,8 @@ where
     async fn transfer_ownership_with_signature(
         &self,
         chain_id: u32,
-        store: Arc<<Self as SubstrateEventWatcher>::Store>,
-        api: Arc<<Self as SubstrateEventWatcher>::Client>,
+        store: Arc<<Self as SubstrateEventWatcher<SubstrateConfig>>::Store>,
+        api: Arc<<Self as SubstrateEventWatcher<SubstrateConfig>>::Client>,
         (public_key, nonce, signature): (Vec<u8>, u32, Vec<u8>),
     ) -> webb_relayer_utils::Result<()> {
         let new_maintainer = public_key.clone();
